requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
docopt==0.6.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
isodate==0.6.0
numpy==1.17.1
oauthlib==3.1.0
pandas==0.25.1
pdfkit==0.6.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
python-slugify==4.0.0